remove Timer and move its methods to Clock
